pyproject.toml
[project]
name = "stockanalysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "akshare>=1.15.0",
    "fastapi>=0.115.0",
    "langchain>=0.2.0",
    "numpy>=1.26.0,<2.0.0",
    "ollama>=0.3.0",
    "pandas>=2.1.0",
    "pyarrow>=15.0.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "requests>=2.31.0",
    "scikit-learn>=1.4.0",
    "torch>=2.0.0",
    "uvicorn>=0.30.0",
    # 认证相关依赖
    "passlib[argon2]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
    "redis>=5.0.1",
    "aioredis>=2.0.1",
    "slowapi>=0.1.8",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.1.0",
]


